% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cars2017.R
\docType{data}
\name{cars2017}
\alias{cars2017}
\title{Fuel Economy Data on Cars in 2017}
\format{A data frame with 1103 observations on the following 7 variables.
\describe{ 
\item{make}{manufacturer of car}
\item{model}{car model}
\item{displacement}{Engine Displacement (Liters)}
\item{cylinders}{number of cylinders}
\item{city_mpg}{fuel economy (mpg), city}
\item{hwy_mpg}{fuel economy (mpg), highway}
\item{gears}{number of gears}
}}
\source{
\url{https://www.fueleconomy.gov}
}
\description{
2017 version of data set for textbook Modern Data
Science With R (see Chapter 9).
}
\keyword{datasets}
