% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deathpenalty.R
\docType{data}
\name{deathpenalty}
\alias{deathpenalty}
\title{Race and the Death Penalty}
\format{
A data frame with 326 observations on the following 3 variables.
\describe{ 
\item{defrace}{Race of the defendant in the capital case}
\item{vicrace}{Race of the victim}
\item{death}{Whether or not the defendant in the case received the death penalty}
}
}
\source{
Michael J. Radelet: "Racial Characteristics and the Imposition of 
the Death Penalty", \emph{American Sociological Review}, 46 (1981).
}
\description{
A dataset recreated from summary data that describes
relationships between race of defendant, race of victim, and outcome 
of trial in a number of capital cases in Florida in 1976-1977.
}
\keyword{datasets}
