% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donation.R
\docType{data}
\name{donation}
\alias{donation}
\title{Zen Center Donations}
\format{
A data frame with 66 observations on the following 5 variables.
\describe{ 
\item{Don_ID}{Donation ID}
\item{Part_ID}{Participant ID}
\item{Don_Amount}{Amount of donation, in dollars.}
\item{Don_Date}{Date of donation.}
\item{Don_Form}{A character vector with two values:  \code{cash} and \code{check}.}
}
}
\source{
Hypothetical data.
}
\description{
Donations made to a fictional Zen Center.  
In a family with \code{\link{participant}}, \code{\link{event}}
and \code{\link{eventParticipation}}.
}
\keyword{datasets}
