% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\docType{data}
\name{event}
\alias{event}
\title{Zen Center Events}
\format{
A data frame with 15 observations on the following 7 variables.
\describe{ 
\item{Event_ID}{Event ID}
\item{Event_Type}{Type of event.  A factor with levels \code{potluck},
\code{retreat} and \code{sangha}.}
\item{Event_Location}{Location of the event.}
\item{Event_Start}{Start-time of the event.}
\item{Event_End}{Ending time of the event.}
\item{Event_Cost}{Nominal cost of the event, in dollars.}
\item{Part_ID}{ID of the participant who organizes the event.}
}
}
\source{
Hypothetical data.
}
\description{
Events at a fictional Zen Center.  
In a family with \code{\link{participant}}, \code{\link{donation}}
and \code{\link{eventParticipation}}.
}
\keyword{datasets}
