% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventParticipation.R
\docType{data}
\name{eventParticipation}
\alias{eventParticipation}
\title{Zen Center Participants}
\format{
A data frame with 107 observations on the following 5 variables.
\describe{ 
\item{Part_ID}{Participant ID}
\item{Event_ID}{ID of the participant.}
\item{EP_Amount_Paid}{Amount actually paid by the participant.}
\item{EP_Notes}{Miscellaneous comments}
\item{EP_Attended}{Whether or not the participant actually attended.}
}
}
\source{
Hypothetical data.
}
\description{
Perticipation of persons in events associated with a
fictional Zen Center.  
In a family with \code{\link{participant}}, \code{\link{donation}}
and \code{\link{event}}.
}
\keyword{datasets}
