% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haliotis.R
\docType{data}
\name{haliotis}
\alias{haliotis}
\title{Age of Sea Snails?}
\format{A data frame with 2923 observations on the following 9 variables.
\describe{ 
\item{Type}{male, femal or infant}
\item{LongSh}{length of longest shell}
\item{Diam}{diameter}
\item{Height}{height}
\item{WhWt}{weight when whole}
\item{ShuckWt}{weight when shucked}
\item{ViscWt}{weight of the snail's viscera}
\item{ShellWt}{weight of the snails's shell}
\item{Age}{number of rings, plus 1.5}
}}
\source{
Dataset derived from the UCI Machine Learning Repository.  See
\url{http://archive.ics.uci.edu/ml/datasets/Abalone} for more information,
including a citation of the original research article.
}
\description{
Subset of data from a study on sea snails of the
genus Haliotis. The age of a such a snail (in years) is quite 
close to the number of rings it has on its shell plus 1.5. The idea
is to predict the age of an individual sea snail from other
characteristics.  The rest of the data is witheld for evalution
purposes.
}
\details{
From the ICU website:  "The age of abalone is determined by 
cutting the shell through the cone, staining it, and counting the number of 
rings through a microscope -- a boring and time-consuming task.  
Other measurements, which are easier to obtain, are used to predict the age."
}
\keyword{datasets}
