% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{parkExp}
\alias{parkExp}
\title{Territoriality:  The Experiment}
\format{A data frame with 237 observations on the following 12 variables.
\describe{ \item{confcar}{The type of car that was waiting for the parking
spot (or that just drove by).  Either a Nissan Maxima or an Infinity Q45. The car is
"confronting" the parked car, hence the
name of the variable.}
\item{sex}{Sex of the driver of the parked car.}
\item{race}{Race of the driver of the parked car.}
\item{num}{Number of people in the parked car (including the driver).}
\item{horn}{Thhe waiting car either honked the horn once, or did
not honk at all.}
\item{carval}{Book value of the parked car, in dollars.}
\item{month}{Month in which the incident occurred.}
\item{day}{Day of the week on which the incident occurred.}
\item{miltime}{Time at which the incident occurred, in military units.
For example, 1130 denotes 11:30AM, while 1350 denotes for 1:50PM.}
\item{time}{Time in seconds for the parked car to depart the parking
spot.}
\item{ccstatus}{Status of the waiting "confronting" car.  The Maxima
is considered a low-status car, whereas the Infinity Q45 is an
expensive, "high-status" car.}
\item{valuediff}{Difference in value between the confronting car and the
parked car, in dollars.  The values of the confronting cars were as follows:
Maxima:  5200, Infinity Q45:  57000.}
 }}
\source{
"Territorial Defense in Parking Lots: Retaliation
Against Waiting Drivers", B. Ruback and D. Juieng, Jounral of Applied
Social Psychology, Volume 27, Issue 9, May 1997, pp. 821-834.
Provided by B. Ruback.
}
\description{
A subset of the \code{parking} data frame, giving only the
subject involved in the experiment.  In the experiment, parked cars
were approached by either an expensive car or a cheap one.  The approaching
car waited for the spot, and while waiting either honked once or did not
honk at all.
}
\note{
This is almost the orginal data.  B. Ruback indicates
(personal communication) that several observations are missing and
cannot be recovered at the present time.
}
\keyword{datasets}

