% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant.R
\docType{data}
\name{participant}
\alias{participant}
\title{Zen Center Participants}
\format{
A data frame with 11 observations on the following 11 variables.
\describe{ 
\item{Part_ID}{Participant ID}
\item{Part_Class}{Type of the participant.  A factor with levels \code{member}
and \code{visitor}.}
\item{Part_Fname}{First name of participant.}
\item{Part_Mname}{Middle name of participant.}
\item{Part_Lname}{Last name of participant.}
\item{Part_Address}{Street address of participant.}
\item{Part_City}{City of participant.}
\item{Part_State}{State of participant.}
\item{Part_Zip}{Postal code of participant.}
\item{Part_Email}{Email address of participant.}
\item{Part_Phone}{Phone number of participant.}
}
}
\source{
Hypothetical data.
}
\description{
Persons associated with fictional Zen Center.  
In a family with \code{\link{event}}, \code{\link{donation}}
and \code{\link{eventParticipation}}.
}
\keyword{datasets}
