% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{payphones}
\alias{payphones}
\title{Territoriality in Pay-Phones}
\format{A data frame with 56 observations on the following 3 variables.
\describe{ \item{sex}{Sex of the subject.}
\item{treatment}{Which condition the subject was put into (A, B
or C as described above) by the researchers.}
\item{time}{Time in seconds that the subject spent on the
phone.}
}}
\source{
R.B. Ruback, K.D. Poe, and P.Doriat, "Waiting on a Phone:
Intrusion on Callers Leads to Territorial Defense" Social Psychology
Quarterly, 52:232-241. Gender data provided by R.B. Ruback (personal
communication).
}
\description{
Will people using a public pay-phone talk longer if someone is
waiting to use their phone? In this experiment, conducted in 1989, "the
investigators
measured the length of time (in seconds) that subjects spent on the telephone
under one of three conditions: when alone (A), when one person was using an
adjacent telphone (B), or when one person was using an adjacent telephone
and another person was waiting to use one of the two telephones. The study
was conducted in an alcove of a shopping mall, an area that contained
only the two adjacent telphones." (Quotation from Business Statistics,
6th. ed. 1992, by W. Daniel and J. Terrell.)
}
\keyword{datasets}

