% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformsNFL.R
\docType{data}
\name{uniformsNFL}
\alias{uniformsNFL}
\title{Malevolence of NFL Uniforms and Penalty Yardage}
\format{
A data frame with 28 observations on the following 3 variables.
\describe{ \item{team}{Name of NFL team.}
\item{malevolence}{Rating of "malevolence" accorded to the eam uniform.  High scores
indicate more malevolence.}
\item{z_pen_yards}{Mean penalty yardage per game for the team, expressed as a z-score.}
}
}
\source{
The Dark Side of Self- and Social Perception:  Black Uniforms and 
Agression in Professional Sports, Frank and Gilovich, Journal of
Personality and Social Psychology, 1988, Vol. 54, No. 1, 74-85.
}
\description{
28 NFL teams from the 1980's.  Team uniforms were rated for their 
"malevolence", and the average penslty yardage for each team was also recorded.
}
\keyword{datasets}
